mirageosPackages.stub-resolver: init
Package a simple MirageOS unikernel as a PoC that this is indeed
possible. Tooling is a bit suboptimal and building an expression that
supports multiple targets is very tedious. However it is probably not
worth it making the situation better before MirageOS 4 is released as
the build system is going to change significantly.

Currently solutions improving quality of life for packaging MirageOS
unikernels would involve import from derivation for downstream usage and
probably code generation for nixpkgs.